+++ tests/data_model_test.py
@@ -42,3 +42,45 @@ def test_experiment_spec_from_and_to_json(beaker_cluster_name):
     }
     spec = ExperimentSpec.from_json(json_spec)
     assert spec.to_json() == json_spec
+
+
+def test_experiment_spec_validation():
+    with pytest.raises(ValidationError, match="Duplicate task name"):
+        ExperimentSpec.from_json(
+            {
+                "tasks": [
+                    {
+                        "name": "main",
+                        "image": {"docker": "hello-world"},
+                        "context": {"cluster": "foo"},
+                        "result": {"path": "/unused"},
+                    },
+                    {
+                        "name": "main",
+                        "image": {"docker": "hello-world"},
+                        "context": {"cluster": "bar"},
+                        "result": {"path": "/unused"},
+                    },
+                ]
+            }
+        )
+    with pytest.raises(ValidationError, match="Duplicate task name"):
+        ExperimentSpec(
+            tasks=[
+                TaskSpec(
+                    name="main",
+                    image={"docker": "hello-world"},
+                    context={"cluster": "foo"},
+                    result={"path": "/unused"},
+                ),
+                TaskSpec(
+                    name="main",
+                    image={"docker": "hello-world"},
+                    context={"cluster": "bar"},
+                    result={"path": "/unused"},
+                ),
+            ]
+        )
+    spec = ExperimentSpec().with_task(TaskSpec.new("main", "foo", docker_image="hello-world"))
+    with pytest.raises(ValueError, match="A task with the name"):
+        spec.with_task(TaskSpec.new("main", "bar", docker_image="hello-world"))